moved most functionality to separate module
